+++ pages/_campus/cars/_id/edit.vue
@@ -65,6 +65,8 @@
         v-if="id"
         type="submit"
         class="button is-primary"
+        :class="{'is-loading': loading}"
+        :disabled="loading"
       >
         <span class="icon is-small">
           <fa-icon :icon="['fas', 'save']" />
@@ -76,6 +78,8 @@
         v-else
         type="submit"
         class="button is-primary"
+        :class="{'is-loading': loading}"
+        :disabled="loading"
       >
         <span class="icon is-small">
           <fa-icon :icon="['fas', 'plus']" />
@@ -90,6 +94,7 @@
 import { mapGetters } from 'vuex';
 import ecField from '~/components/form/field.vue';
 import SearchCarModels from '~/components/form/search-car-models.vue';
+import toggleLoading from '~/helpers/mixins/toggle-loading';
 
 const EDITABLE_FIELDS = ['id', 'label', 'model'];
 
@@ -98,6 +103,7 @@ export default {
     SearchCarModels,
     ecField,
   },
+  mixins: [toggleLoading],
   props: {
     car: {
       type: Object,
@@ -119,6 +125,7 @@ export default {
     async edit(car) {
       let data = {};
       try {
+        this.toggleLoading(true);
         if (this.id) {
           ({ data } = (await this.CarsAPI.patchCar(car.id, car)));
         } else {
@@ -135,6 +142,7 @@ export default {
           this.$toast.error('Erreur serveur, si le problème persiste, veuillez contacter le service technique');
         }
       }
+      this.toggleLoading(false);
     },
   },
 };