+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 ## Current
 ### Features
 ### Fix
+## Version 1.4.6
+### Features
+* Added loading animation and success message while creating or editing CRUD
+* Added alert message if browser is not Firefox
+### Fix
+* Add missing create button in Phone index
+* Add missing create button in POI index
+* Add missing create button in users index
+* Reservation scope can be set in days on base creation
+* Status color in page history
+* Fixed status update in page history
+* Fixed driver name display in history table
+* Fix SSR rendering of pages where calendar is present
+* Fixed rights to import csv and create new data for phones and pois
+* Fixed csv export mask to fit import template
+* Added format-coordinates.js to generate number compatible format
 ## Version 1.4.5
 ### Features
 * Add driver connection status in planificator